Simplify printing to console
Previously, printing output to console was abstracted. This caused issues when text was printed to the very bottom line of the console, causing the console to scroll, and all the hardcoded row positions to be offset from their intended location. This resulted in text overwriting the wrong location and erroneous output. Simply using carriage return fixes all these issues.